feat: add order model

+++ db/migrate/20250830183112_create_products.rb
@@ -3,7 +3,7 @@
 class CreateProducts < ActiveRecord::Migration[8.0]
   def change
     create_table :products do |t|
-      t.integer :external_id, index: { unique: true }, null: false
+      t.bigint :external_id, index: { unique: true }, null: false
       t.integer :price_cents, null: false
 
       t.timestamps
